Added files via upload
The files were mixed up since I still don't know how to have files uploaded in here with organization. I hope you'll consider. There are three folders in MatrixUserBundle ,  the controller, entity and resources inside the controller are the files with controller in their file name and the .twig are inside the resources and the the user.php is inside the entity. I hope you understand. If you have some question feel free to ask. Thanks again!